transformer-proxy:code cleanup, adopting factory for proxy layer
