--- Dockerfile-proxy
@@ -1,21 +1,0 @@
-FROM node:10.16.0-alpine
-
-RUN apk add --no-cache tini python make g++
-RUN mkdir -p /home/node/app/node_modules && chown -R node:node /home/node/app
-
-# Create app directory
-WORKDIR /home/node/app
-USER node
-
-ARG version
-ENV transformer_build_version=$version
-COPY package*.json ./
-RUN npm install
-
-COPY --chown=node:node . .
-
-ENTRYPOINT ["/sbin/tini", "--"]
-CMD [ "npm", "run", "proxy" ]
-
-
-EXPOSE 9091/tcp
