+++ package.json
@@ -20,7 +20,6 @@
     "lint-json-fix": "npx eslint --ext .json --fix .",
     "merge-check": "npm run verify || exit 1; codecov",
     "start": "node destTransformer.js",
-    "proxy": "node proxy.js",
     "testDestinations": "npx jest --runInBand --detectOpenHandles --coverage --notify",
     "testVersionedRouter": "npx jest versionedRouter --type=all --detectOpenHandles --notify",
     "testVersioned": "npx jest versionedRouter --destName=heap --type=all --notify",
