+++ v0/destinations/marketo/util.js
@@ -2,7 +2,8 @@ const { httpGET, httpPOST } = require("../../../adapters/network");
 const {
   nodeSysErrorToStatus,
   trimResponse,
-  getDynamicMeta
+  getDynamicMeta,
+  processAxiosResponse
 } = require("../../../adapters/utils/networkUtils");
 const { isHttpStatusSuccess } = require("../../util/index");
 const { TRANSFORMER_METRIC } = require("../../util/constant");
