+++ v0/destinations/trengo/transform.js
@@ -1,10 +1,10 @@
 /* eslint-disable no-nested-ternary */
 /* eslint-disable no-prototype-builtins */
 const Handlebars = require("handlebars");
+const axios = require("axios");
 const get = require("get-value");
 const { EventType } = require("../../../constants");
 const { EndPoints, BASE_URL } = require("./config");
-const { httpGET } = require("../../../adapters/network");
 const {
   getHashFromArray,
   removeUndefinedAndNullValues
@@ -81,7 +81,7 @@ const validate = (email, phone, channelIdentifier) => {
 const lookupContact = async (term, destination) => {
   let res;
   try {
-    res = await httpGET(`${BASE_URL}/contacts?page=1&term=${term}`, {
+    res = await axios.get(`${BASE_URL}/contacts?page=1&term=${term}`, {
       headers: {
         Authorization: `Bearer ${destination.Config.apiToken}`
       }
